discord_bot/requirements.txt
aiohttp
aiosqlite
discord.py==2.6.4
google-generativeai
python-dotenv